backend/requirements.txt
langchain==0.1.0
langchain_community
langchain-core
langgraph==0.0.24
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0